Working stats pre cog

+++ dbLayer/app/app/models.py
@@ -38,11 +38,11 @@ class Beat(db.Model):
 class Prediction(db.Model):
 	_tablename_ = 'prediction'
 	ID = db.Column(db.Integer, primary_key = True)
-	certainty = db.Column(db.Integer)
+	certainty = db.Column(db.Float)
 	location = db.Column(Geometry('POINT'))
 	type = db.Column(db.String(132))
-	precog = db.Column(db.String(16)) # Who made the prediction
-	datetime = db.Column(db.DateTime) # When was the prediction made
+	precog = db.Column(db.String(16))
+	datetime = db.Column(db.DateTime)
 
 class Recommendation(db.Model):
 	_tablename_ = 'recommendation'
